Split applications into different files
