+++ test/cloudtest/test_cluster_rehydration.py
@@ -9,7 +9,7 @@
 
 from textwrap import dedent
 
-from materialize.cloudtest.application import MaterializeApplication
+from materialize.cloudtest.materialize_application import MaterializeApplication
 
 
 # Test that a crashed (and restarted) cluster replica handles rehydration
