+++ test/cloudtest/test_metrics.py
@@ -9,7 +9,7 @@
 
 from textwrap import dedent
 
-from materialize.cloudtest.application import MaterializeApplication
+from materialize.cloudtest.materialize_application import MaterializeApplication
 
 
 def test_replica_metrics(mz: MaterializeApplication) -> None:
