+++ test/cloudtest/test_storage.py
@@ -12,9 +12,9 @@ from textwrap import dedent
 import pytest
 from pg8000.exceptions import InterfaceError
 
-from materialize.cloudtest.application import MaterializeApplication
 from materialize.cloudtest.exists import exists, not_exists
 from materialize.cloudtest.k8s import cluster_pod_name, cluster_service_name
+from materialize.cloudtest.materialize_application import MaterializeApplication
 from materialize.cloudtest.wait import wait
